02_frameworks/5_crewai/code/starter_project/pyproject.toml
[project]
name = "starter_project"
version = "0.1.0"
description = "A starter project demonstrating CrewAI capabilities"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.11.0",
    "python-dotenv>=0.19.0",
    "openai>=1.0.0",
    "crewai-tools>=0.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py310" 